(MODULES-4314) Standardize using ModuleSync
This commit updates the components standardized by running
```
msync update --noop -f puppetlabs-certregen
```
against the project.

Prior to this commit, the Jenkins testing pipeline could not properly
install the gem bundle on the test runner with the committed
Gemfile.lock. Applying this update brings the module components
into a state that can be maintained in the future by adding the
module to the ModuleSync list of managed modules.