(MODULES-4317) Test redistribute - ensure setup
This commit ensures that the initial setup of distributing a CA that is
updated after expiration succeeds.

Prior to this commit, the CA might have expired before a puppet run updated
it on the nodes. Now, a successful puppet run to set up the short ttl cert
on the nodes is required. An additional validation to ensure that the master
can ssh to the agents has been added.