(MODULES-4317) workflow acceptance: fix ssh key
This commit changes the ssh-keygen on the master node to use the
standard `id-rsa` keyname and ensures that the agent hostnames are
added to the `known_hosts` file by setting `StrictHostKeyChecking=no`.
This is done to ensure that the `certregen redistribute` command does
not have any difficulty performing ssh commands on the agents.